requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.10.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
gradio==4.44.1
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
unstructured>=0.10.0

statsmodels>=0.14.0
scikit-learn>=1.3.0
seaborn>=0.13.0
audioop-lts>=0.2.0
